requirements.txt
appdirs==1.4.3
asgiref==3.2.7
beautifulsoup4==4.8.2
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
configparser==5.0.0
cycler==0.10.0
distlib==0.3.0
Django==3.0.4
djangorestframework==3.11.0
djangorestframework-filters==0.11.1
docopt==0.6.2
filelock==3.0.12
googletrans==2.4.0
httpie==2.0.0
idna==2.9
Js2Py==0.68
kiwisolver==1.1.0
Markdown==3.2.1
matplotlib==3.2.1
mysqlclient==1.4.6
nltk==3.4.5
numpy==1.18.2
packaging==20.3
pbs==0.110
pipwin==0.4.8
PyAudio==0.2.11
Pygments==2.6.1
pyjsparser==2.7.1
pyparsing==2.4.6
PyPrind==2.11.2
pySmartDL==1.3.3
python-dateutil==2.8.1
pytz==2019.3
requests==2.23.0
six==1.14.0
soupsieve==2.0
SpeechRecognition==3.8.1
sqlparse==0.3.1
stdio==0.10
tzlocal==2.0.0
urllib3==1.25.8
virtualenv==20.0.15
